Improved/fixed some docstrings.  E.g. one said a list of analyses was returned, but I'd changed to return singles.
Fixed a bug where the "unlabeled-malware" label was added as a
string instead of a one-element list of strings.

Restructured some analysis translation code to ensure that
returned STIX analyses always had a "results" property.  That
property is required in STIX, but it was possible for it to be
missing in the translation.  I had a comment because I wasn't
sure at the time about what to do... but I decided that if a
"results" property couldn't be obtained, to fail the translation
and not produce a STIX analysis.